+++ lib/guard/runner.rb
@@ -5,11 +5,11 @@ require 'guard/watcher'
 
 module Guard
 
-  # The runner is responsible for running all methods defined on each guards.
+  # The runner is responsible for running all methods defined on each plugin.
   #
   class Runner
 
-    # Runs a Guard-task on all registered guards.
+    # Runs a Guard-task on all registered plugins.
     #
     # @param [Symbol] task the task to run
     # @param [Hash] scopes either the Guard plugin or the group to run the task on
@@ -18,7 +18,7 @@ module Guard
     #
     def run(task, scopes = {})
       Lumberjack.unit_of_work do
-        _scoped_guards(scopes) do |guard|
+        _scoped_plugins(scopes) do |guard|
           run_supervised_task(guard, task) if guard.respond_to?(task)
         end
       end
@@ -28,7 +28,7 @@ module Guard
     ADDITION_TASKS     = [:run_on_additions, :run_on_changes, :run_on_change]
     REMOVAL_TASKS      = [:run_on_removals, :run_on_changes, :run_on_deletion]
 
-    # Runs the appropriate tasks on all registered guards
+    # Runs the appropriate tasks on all registered plugins
     # based on the passed changes.
     #
     # @param [Array<String>] modified the modified paths.
@@ -37,7 +37,7 @@ module Guard
     #
     def run_on_changes(modified, added, removed)
       ::Guard::UI.clearable
-      _scoped_guards do |guard|
+      _scoped_plugins do |guard|
         modified_paths = ::Guard::Watcher.match_files(guard, modified)
         added_paths    = ::Guard::Watcher.match_files(guard, added)
         removed_paths  = ::Guard::Watcher.match_files(guard, removed)
@@ -73,7 +73,7 @@ module Guard
         ::Guard::UI.error("#{ guard.class.name } failed to achieve its <#{ task.to_s }>, exception was:" +
                  "\n#{ ex.class }: #{ ex.message }\n#{ ex.backtrace.join("\n") }")
 
-        ::Guard.guards.delete guard
+        ::Guard.plugins.delete guard
         ::Guard::UI.info("\n#{ guard.class.name } has just been fired")
 
         ex
@@ -85,7 +85,7 @@ module Guard
     # @note If a Guard group is being run and it has the `:halt_on_fail`
     #   option set, this method returns :no_catch as it will be caught at the
     #   group level.
-    # @see ._scoped_guards
+    # @see ._scoped_plugins
     #
     # @param [Guard::Plugin] guard the Guard plugin to execute
     # @return [Symbol] the symbol to catch
@@ -126,16 +126,16 @@ module Guard
     # @param [Hash] scopes hash with plugins or a groups scope
     # @yield the task to run
     #
-    def _scoped_guards(scopes = {})
-      if guards = _current_plugins_scope(scopes)
-        guards.each do |guard|
+    def _scoped_plugins(scopes = {})
+      if plugins = _current_plugins_scope(scopes)
+        plugins.each do |guard|
           yield(guard)
         end
       else
         _current_groups_scope(scopes).each do |group|
           current_plugin = nil
           block_return = catch :task_has_failed do
-            Array(::Guard.guards(:group => group.name)).each do |guard|
+            Array(::Guard.plugins(:group => group.name)).each do |guard|
               current_plugin = guard
               yield(guard)
             end